Feat: 불필요한 콘솔 제거
